refactor(sdk): extract manifest package out of cli/register (fix layering leak)
