+++ hexgate/cli/register/__init__.py
@@ -1,8 +1,11 @@
-"""Agent registration: Python API and `hexgate register` CLI."""
+"""The `hexgate register` CLI + platform-registration API.
 
-from hexgate.cli.register.register import register_agent
-from hexgate.cli.register.manifest import create_manifest
-from hexgate.cli.register.models import AgentManifest
+Manifest models and the builder now live in :mod:`hexgate.manifest`;
+this package only carries platform registration (`register_agent`,
+`post_manifest`) and the CLI entry point.
+"""
+
 from hexgate.cli.register.main import add_parser, main
+from hexgate.cli.register.register import register_agent
 
-__all__ = ["register_agent", "create_manifest", "AgentManifest", "add_parser", "main"]
+__all__ = ["register_agent", "add_parser", "main"]
