+++ hexgate/manifest/openai.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from hexgate.cli.register.models import (
+from hexgate.manifest.models import (
     AgentFramework,
     AgentManifest,
     InputProperty,
