+++ tests/register/test_manifest.py
@@ -1,16 +1,16 @@
-from hexgate.cli.register.manifest import create_manifest
-from hexgate.cli.register.models import (
+from hexgate.manifest import create_manifest
+from hexgate.manifest.models import (
     AgentManifest,
     AgentFramework,
     InputProperty,
     InputSchema,
     ToolDefinition,
 )
-from hexgate.cli.register.hexgate import create_hexgate_manifest
-from hexgate.cli.register.openai import create_openai_manifest
-from hexgate.cli.register.google import create_google_manifest
-from hexgate.cli.register.langchain import create_langchain_manifest
-from hexgate.cli.register.pydantic_ai import create_pydantic_ai_manifest
+from hexgate.manifest.native import create_hexgate_manifest
+from hexgate.manifest.openai import create_openai_manifest
+from hexgate.manifest.google import create_google_manifest
+from hexgate.manifest.langchain import create_langchain_manifest
+from hexgate.manifest.pydantic_ai import create_pydantic_ai_manifest
 
 
 def test_agent_manifest_schema():
